requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
gspread>=6.1.0
google-auth>=2.29.0
openpyxl>=3.1.2
